backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
filetype==1.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.0
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.3.21
langchain-google-genai==2.0.6
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.147
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.12
packaging==24.2
# pinecone==5.4.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pip==24.2
propcache==0.2.1
proto-plus==1.25.0
# protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
pydantic-settings==2.6.1
Pygments==2.18.0
pyparsing==3.2.0
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.3
tenacity==9.0.0
tqdm==4.67.1
typer==0.14.0
typing_extensions==4.12.2
typing-inspect==0.9.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
wheel==0.44.0
yarl==1.18.3
